requirements.txt
# Django
Django>=2.1.10
# postgresql middleware
# psycopg2==2.7.6.1 --no-binary psycopg2
# psycopg2==2.7.6.1
# psycopg2-binary==2.7.6.1
# django-redis and redis
django-redis==4.10.0
redis==3.2.0
# Restful API
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
django-cors-headers==3.0.2
# job queues and job scheduling
celery==4.3.0

# utils
python-dotenv==0.10.1
requests==2.21.0
mistune==0.8.3
xpinyin==0.5.6
shortuuid==0.5.0

# development only
django-extensions==2.1.4
django-debug-toolbar==1.11
better-exceptions==0.2.2

# production
gunicorn==19.9.0
